doclava17: Add SourcePosition stub implementation
Current progress is 126/147 methods implemented (85.71%).

Source position info is used only for reporting warnings and
errors while producing documentation and it does not end up in the
generated docs. In conjunction with the lack of the source position
info in the DocletEnvironment, there is no easy way of getting that,
so the current implementation is a stub one.

If required, this could be implemented with a separate compilation
pass using com.sun.source.util.JavacTask (for source/binary). Another
approach is to use ANTLR to parse sources (this was started by
previous doclava maintainers in src/com/google/doclava/parser).

Bug: 260694901
Test: ./gradlew doclava17:test
Change-Id: I0f8ff191dd8b679f31122b70d9a9e959bf199394