Make doclava support baselining of lint issues
There are many existing issues in the docs. Before new errors can
be forbidden, make it possible to allow pre-existing errors.

The baseline support is very crude -- any mention of a file will allow
all lint errors in that file. It's my hope to trim the baseline down
quickly so that this won't be an issue.

Bug: 217734059
Test: in follow-up changes which enables errors and passes baselines
Change-Id: I78174f66f9d5cd59511951a481072f54339eba11