Avoid exception if comment ends with `@`
Previously, it would throw an exception when processing a comment like
this.
```
    /**
     * ....
     * @return true if number contains @
     */
```

This was not an issue before because doclava was ignoring all
`@SystemApi` annotated APIs because while Metalava removed the
`@SystemApi` annotations it left the `@hide`. That meant that doclava
just ignored them. When that Metalava bug was fixed these problems were
revealed.

Bug: 462333324
Test: m offline-system-sdk-referenceonly-docs
Change-Id: I0b76b183910a4397bd1229b84bf5ba47f95540c1